refactor all codes
