+++ components/organisms/tip/New.vue
@@ -56,7 +56,7 @@ import Button from '~/components/atoms/Button'
 import InputForm from '~/components/atoms/InputForm'
 import SingleSelectForm from '~/components/atoms/SingleSelectForm'
 import MultipleSelectForm from '~/components/atoms/MultipleSelectForm'
-import { CATEGORIES } from '~/utils/categories'
+import { CATEGORIES } from '~/utils/index'
 export default {
   components: {
     FormTemplate,
@@ -78,7 +78,7 @@ export default {
   },
   computed: {
     categories () {
-      return CATEGORIES.map(category => {return category.name})
+      return CATEGORIES
     },
     ...mapGetters(['events'])
   },
