+++ pages/flight/index.vue
@@ -7,6 +7,18 @@
     <Status
       :list="flights.item"
     />
+    <SingleSelectForm
+      :option="years"
+      :number="params.year"
+      column="年"
+      @form-data="applyYear"
+    />
+    <SingleSelectForm
+      :option="boardingTypes"
+      :number="params.boardingType"
+      column="搭乗機材"
+      @form-data="applyBoardingType"
+    />
     <FlightList
       :list="flights.item"
       :number="params.page"
@@ -30,15 +42,18 @@ import moment from 'moment'
 import { mapGetters, mapState } from 'vuex'
 import MainTemplate from '~/components/templates/MainTemplate';
 import Pagination from '~/components/atoms/Pagination'
+import SingleSelectForm from '~/components/atoms/SingleSelectForm'
 import FlightList from '~/components/organisms/flight/List'
 import NewFlight from '~/components/organisms/flight/New'
 import EditFlight from '~/components/organisms/flight/Edit'
 import Status from '~/components/organisms/flight/Status'
+import { BOARDING_TYPE_LIST, YEARS } from '~/utils/index'
 export default {
   middleware: 'auth',
   components: {
     MainTemplate,
     Pagination,
+    SingleSelectForm,
     FlightList,
     NewFlight,
     EditFlight,
@@ -56,12 +71,19 @@ export default {
       },
       dataKey: '',
       params: {
-        page: 1
-      }
+        page: 1,
+        year: 0,
+        boardingType: 0
+      },
+      boardingTypes: BOARDING_TYPE_LIST,
+      years: YEARS
     }
   },
   async mounted() {
-    this.flights ? Promise.resolve() : this.$store.dispatch('initFlights')
+    await this.$store.dispatch('initFlights', {
+      boardingType: this.params.boardingType,
+      year: this.params.year
+    })
   },
   computed: {
     ...mapGetters(['userStatus', 'flights']),
@@ -78,6 +100,20 @@ export default {
     },
     applyPage(value) {
       this.params.page = value
+    },
+    async applyBoardingType(value) {
+      this.params.boardingType = value
+      await this.$store.dispatch('initFlights', {
+        boardingType: this.params.boardingType,
+        year: this.params.year
+      })
+    },
+    async applyYear(value) {
+      this.params.year = value
+      await this.$store.dispatch('initFlights', {
+        boardingType: this.params.boardingType,
+        year: this.params.year
+      })
     }
   }
 }
