--- utils/categories.js
@@ -1,36 +1,0 @@
-export const CATEGORIES = [
-  {value: 1, text: 'Web', backgroundColor: '#0033ff', color: '#fff'},
-  {value: 2, text: 'HTML5', backgroundColor: '#0099ff', color: '#000'},
-  {value: 3, text: 'CSS', backgroundColor: '#00ff99', color: '#000'},
-  {value: 4, text: 'Javascript', backgroundColor: '#ffff00', color: '#000'},
-  {value: 5, text: 'EcmaScript', backgroundColor: '#ff9966', color: '#000'},
-  {value: 6, text: 'PHP', backgroundColor: '#000099', color: '#fff'},
-  {value: 7, text: 'Zendframework', backgroundColor: '#009933', color: '#fff'},
-  {value: 8, text: 'Laravel', backgroundColor: '#ff99cc', color: '#000'},
-  {value: 9, text: 'Scala', backgroundColor: '#ff0000', color: '#fff'},
-  {value: 10, text: 'Playframework', backgroundColor: '#666666', color: '#fff'},
-  {value: 11, text: 'Kotlin', backgroundColor: '#ffcc33', color: '#000'},
-  {value: 12, text: 'Sparkframework', backgroundColor: '#cccccc', color: '#000'},
-  {value: 13, text: 'Heroku', backgroundColor: '#ccffff', color: '#000'},
-  {value: 14, text: 'Netlify', backgroundColor: '#99ffcc', color: '#000'},
-  {value: 15, text: 'AWS', backgroundColor: '#660000', color: '#fff'},
-  {value: 16, text: 'React', backgroundColor: '#3300ff', color: '#fff'},
-  {value: 17, text: 'Vue', backgroundColor: '#33ff33', color: '#000'},
-  {value: 18, text: 'Objective-c', backgroundColor: '#cc9933', color: '#000'},
-  {value: 19, text: 'Swift', backgroundColor: '#ff6633', color: '#fff'},
-  {value: 20, text: 'Firebase', backgroundColor: '#ff9900', color: '#fff'},
-  {value: 21, text: 'TypeScript', backgroundColor: '#ff33ff', color: '#fff'},
-  {value: 22, text: 'Nuxt.js', backgroundColor: '#339933', color: '#fff'},
-  {value: 23, text: 'VuePress', backgroundColor: '#33ff99', color: '#000'},
-  {value: 24, text: 'Gatsby', backgroundColor: '#990099', color: '#fff'},
-  {value: 25, text: 'Angular', backgroundColor: '#ff3333', color: '#fff'},
-  {value: 26, text: 'Elixir', backgroundColor: '#660099', color: '#fff'},
-  {value: 27, text: 'Java', backgroundColor: '#ff6600', color: '#fff'},
-  {value: 28, text: 'Docker', backgroundColor: '#1D2088', color: '#fff'},
-  {value: 29, text: 'MySQL', backgroundColor: '#ff9933', color: '#fff'},
-  {value: 30, text: 'PostgreSQL', backgroundColor: '#333399', color: '#fff'},
-  {value: 31, text: 'Redis', backgroundColor: '#cc0000', color: '#fff'},
-  {value: 99, text: 'Others', backgroundColor: '#000000', color: '#fff'}
-]
-
-export default {}
